Refactor: put fragments and activities and solver in their own respective packages
