+++ app/src/main/res/layout/activity_start_screen.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#0099cc"
-    tools:context="com.captainhampton.android.lightsout.StartScreenActivity">
+    tools:context=".activity.ActivityStartScreen">
 
     <com.pddstudio.starview.StarView
         android:layout_width="match_parent"
