+++ app/src/main/res/layout/fragment_menu.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.captainhampton.android.lightsout.Main">
+    tools:context=".activity.ActivityMain">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
